requirements.txt
pyannote.audio==1.1
torch==-1.9.0